Server and webpage changes. Added features and security.
Lots of changes here. Refactoring of the server code.
Uses Client to hold the data from each client. Username, colors, etc is
held on the server and reassigned every time a client reconnects to the
server. Provides abstraction for sessions and etc.
GroupManager manages the groups and which client connects to each group.
Chat_color enumarator is a list of allowed chat colors. Group Manager
reserves 'red' for TA_X usernames.
Wrote small functions for converting xmls and sendingMessages.
Rewrote webpages and server validation so user's cannot skip landing
page. If clients do not have a valid ID they are sent back to the
landing page (might allow login in later update).
Added a lot of features for each page. 
Admin displays confirmation if groups already exists, and shows how many
groups exist.
Login will automatically update with the number of groups, and will lock
users if no groups exist.
Chat has different backgrounds to show if the message is from chat
history, the user is typing, or the message is from yourself. A list of
group members is on the right. Chat also pulls chatHistory for late
group members and displays which group the user is in.
FileLogger now logs all chat messages that have been broadcast to the
group. This includes joinGroup, typing, chat and endGroup messages.