Refining Answer system & bugfixes
Changes to the Answer system to provide more feedback.
Timers for the answer lock.
Answer Popup to alert the group of submission status
Answer window now has 1 window, locked bar changed to be general status
Drop down on student page showing for status of the answer

Various bugfixes on Answer and other systems