Added IP address in messages
IP address is now read on the client side using ipify and is 
transmitted with every message sent from loginChatJS.js to then
be saved in the log.

This update does not include an IP address with messages that 
are rebroadcast to the group. This functionality will need to
be added in ChatAnnotation.java.