NIGHTLY COMMIT 10-25-17
Still working on HTTP Request handler. have Fileloggers writing file
names to currentLogFileNames.txt and request handler is reading the
filenames successfully, but when I try to parse the xml file it is
returning a null value. 

Added a cheat to read from the file path, but need to fix so that it is
getting the logPath from web.xml